refactor(core): define usage page as uint16_t
Aligns with the HID specification.

Usage page values were sometimes declared as uint8_t and sometimes uint16_t.  This commit aligns all instances with the HID specification for consistency.

PR: #521

+++ app/include/zmk/events/keycode_state_changed.h
@@ -14,7 +14,7 @@
 
 struct keycode_state_changed {
     struct zmk_event_header header;
-    uint8_t usage_page;
+    uint16_t usage_page;
     uint32_t keycode;
     uint8_t implicit_modifiers;
     bool state;
